implemented ArgumentInformation, rearranged code
Implemented the ArgumentInformation class to replace the Strings that
would contain argument names. This allows us to easily store both
argument names and description neat and clean in one object.

Added get and set methods for "programDescription" which would contain
the program description (like the example screenshot in the groupme)

Commented out the method that added arguments using an array of strings,
as it was broken with the new class, and we need to find an way to
re-implement that.

Rearranged some code so that relevant things are closer together.

Also added/updated all appropriate unit tests.