Changed error messages. Made two more demos.
Error messages for the TooMany and NotEnough arguments now state if it
was too many or too few and also the amount given versus the amount
expected on top of the list.

There are three demos now. One that passes, one that fails with too many
arguments and one that fails with not enough arguments.